requirements.txt
simpy
numpy
kafka-python
pyspark
alibi-detect
mlflow
fastapi
uvicorn
streamlit
requests
torch
streamlit-autorefresh